// api/models/paginated-vendors.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PardoMart Node API
 * API documentation for the PardoMart Node.js application.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@pardomart.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { VendorListItem } from './vendor-list-item';

export interface PaginatedVendors {
    'page'?: number;
    'totalPages'?: number;
    'pageSize'?: number;
    'totalCount'?: number;
    'data'?: Array<VendorListItem>;
}